2-ApiMobileBasico/requirements.txt
robotframework==4.1.2
robotframework-seleniumlibrary
datetime
dateutils
robotframework-requests
robotframework-jsonlibrary
robotframework-appiumlibrary
robotframework-pabot
allure-robotframework
RESTinstance